logs, clearer structure, updated interface
