+++ conn.go
@@ -1,12 +1,7 @@
 package minigo
 
 import (
-	"fmt"
-	"log"
-	"strings"
 	"time"
-
-	"go.bug.st/serial"
 )
 
 type Connector interface {
@@ -14,148 +9,9 @@ type Connector interface {
 
 	Write([]byte) error
 
-	Read() (int, []byte, error)
+	Read() ([]byte, error)
 
-	ReadTimeout(time.Duration) (int, []byte, error)
+	ReadTimeout(time.Duration) ([]byte, error)
 
 	Connected() bool
-}
-
-type Modem struct {
-	port        serial.Port
-	init        []ATCommand
-	buffer      []byte
-	ringHandler func(modem *Modem)
-	connected   bool
-}
-
-type ATCommand struct {
-	Command string
-	Reply   string
-}
-
-func NewModem(portName string, baud int, init []ATCommand) *Modem {
-	port, err := serial.Open(portName, &serial.Mode{BaudRate: baud})
-	if err != nil {
-		errorLog.Fatalf("unable to start modem port=%s baud=%d: %s", port, baud, err.Error())
-	}
-
-	return &Modem{
-		port:      port,
-		init:      init,
-		buffer:    make([]byte, 1024),
-		connected: false,
-	}
-}
-
-func (m *Modem) Init() error {
-	for _, at := range m.init {
-		if !m.sendCommandAndWait(at) {
-			return fmt.Errorf("cannot ack command='%s'", at.Command)
-		}
-	}
-
-	return nil
-}
-
-func (m *Modem) sendCommandAndWait(at ATCommand) bool {
-	// Send initial message
-	if len(at.Command) > 0 {
-		if _, err := m.port.Write([]byte(at.Command + "\r\n")); err != nil {
-			log.Println(err)
-		}
-	}
-
-	ack := false
-	// Wait for message
-	if len(at.Reply) > 0 {
-		var result string
-		for {
-			n, buffer, err := m.ReadTimeout(5 * time.Second)
-			if err != nil {
-				errorLog.Println(err)
-				break
-			}
-			if n == 0 {
-				break
-			}
-
-			result += string(buffer[0:n])
-			if strings.Contains(result, at.Reply) {
-				ack = true
-				break
-			} else if strings.Contains(result, "ERROR") {
-				errorLog.Println(err)
-				break
-			}
-		}
-	} else {
-		ack = true
-	}
-
-	return ack
-}
-
-func (m *Modem) Write(b []byte) error {
-	_, err := m.port.Write(b)
-	return err
-}
-
-func (m *Modem) Read() (int, []byte, error) {
-	n, err := m.port.Read(m.buffer)
-	return n, m.buffer, err
-}
-
-func (m *Modem) ReadTimeout(d time.Duration) (int, []byte, error) {
-	m.port.SetReadTimeout(d)
-	defer m.port.SetReadTimeout(serial.NoTimeout)
-
-	n, err := m.port.Read(m.buffer)
-	return n, m.buffer, err
-}
-
-func (m *Modem) Connected() bool {
-	return m.connected
-}
-
-func (m *Modem) RingHandler(f func(modem *Modem)) {
-	m.ringHandler = f
-}
-
-func (m *Modem) Serve(forceRing bool) {
-	var err error
-	var status *serial.ModemStatusBits
-
-	for {
-		status, err = m.port.GetModemStatusBits()
-		if err != nil {
-			warnLog.Printf("unable to get modem status: %s\n", err.Error())
-		}
-
-		if !status.DCD && m.connected {
-			infoLog.Printf("closed connection\n")
-			m.connected = false
-			m.Init()
-		}
-
-		if status.RI || forceRing {
-			infoLog.Printf("phone rings\n")
-			forceRing = false
-			m.Connect()
-		}
-
-		time.Sleep(time.Second)
-	}
-
-}
-
-func (m *Modem) Connect() {
-	if !m.sendCommandAndWait(ATCommand{Command: "ATA", Reply: "CONNECT 1200/75/NONE"}) {
-		errorLog.Printf("unable to connect after Ring")
-		return
-	}
-	m.connected = true
-	infoLog.Printf("connection established\n")
-
-	go m.ringHandler(m)
 }
